cloud-threat-detection-platform/backend/requirements.txt
psutil
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
kafka-python
python-multipart
requests
python-dotenv
scikit-learn
numpy
joblib
jinja2
slowapi
gunicorn